requirements.txt
Flask==3.0.2
Werkzeug==3.0.1
scikit-learn==1.6.1
statsmodels==0.14.1
xgboost==2.0.3
lifelines==0.27.8
flask-login==0.6.3
flask-sqlalchemy==3.1.1
flask-wtf==1.2.1
flask-mail==0.9.1
flask-migrate==4.1.0
email-validator==2.1.0.post1
numpy==1.26.4
pandas==2.2.1
python-dotenv==1.0.1
gunicorn==21.2.0
python-docx==1.1.0
reportlab==4.1.0
transformers==4.39.3
requests==2.31.0
openai==1.10.0
psycopg2-binary==2.9.9
matplotlib==3.8.3
scipy==1.12.0
seaborn==0.13.1
pillow==10.2.0
itsdangerous==2.1.2
# Required for rendering plots in web interface
plotly==5.18.0
# Required for Hugging Face models
torch==2.1.2
